feat: implement post likes in community screen
Adds the functionality to like posts in the community screen.
Users can now like posts, and the number of likes will increment.
The like operation will update the like field in firestore by incrementing the value by 1.
It also add onLike and onComment callbacks in the community screen and community post elements.